--- frontend/zona44app/lib/pages/Perfil/Admin/OrderAdmin/views/OrderDetails/cubit/order_details_cubit.dart
@@ -1,8 +1,0 @@
-import 'package:bloc/bloc.dart';
-import 'package:equatable/equatable.dart';
-
-part 'order_details_state.dart';
-
-class OrderDetailsCubit extends Cubit<OrderDetailsState> {
-  OrderDetailsCubit() : super(OrderDetailsInitial());
-}
