--- frontend/zona44app/lib/pages/Perfil/Admin/OrderAdmin/views/OrderDetails/cubit/order_details_state.dart
@@ -1,10 +1,0 @@
-part of 'order_details_cubit.dart';
-
-sealed class OrderDetailsState extends Equatable {
-  const OrderDetailsState();
-
-  @override
-  List<Object> get props => [];
-}
-
-final class OrderDetailsInitial extends OrderDetailsState {}
